requirements.txt
streamlit
Pillow
azure-cognitiveservices-speech
pydub
